naive forecaster
This is the benchmark forecast.

For market questions, it forecasts the market value on T-1 (for forecasts made on day T).
For dataset questions, it calculates a baserate for the question and uses that as its forecast.

Baserate calculations are basic for ACLED at the moment. Should be improved and created for all
other dataset questions.