infer: fix bug
The previous way of determining binary questions was not correct.

It obtained questions that were of non-binary types or of type "continuous binary" but we cannot resolve these correctly.

The possible types are:
[
'Forecast::Binary::ContinuousQuestion',
'Forecast::ContinuousYesNoQuestion',
'Forecast::NonExclusiveOpinionPoolQuestion',
'Forecast::Question',
'Forecast::YesNoQuestion'
]

Of these, the only type we're interested in are of type `'Forecast::YesNoQuestion'`