metaculus: update code based on best guesses following undocumented API changes
See https://www.metaculus.com/notebooks/28595/104-update-updates-to-metaculus-api/ for background

The `api2/questions/` endpoint used for fetching new questions does not handle the parameters we
passed correctly and nothing was mentioned about it in the blogpost. Hence, stop fetching and update
the code when the API is documented.

The `api2/questions/{market_id}` endpoint works given the changes mentioned in the blogpost and the
assumption that `recency_weighted` corresponds to the former `community_prediction`. Though they're
different, it seems the differences track on the graphs displayed on the website as well, implying
Metaculus has changed the way they calculate the community prediction.