fred: nullify AMERIBOR
FRED only maintains data on `AMERIBOR` for two weeks; all preceding data is deleted. This currently
poses a problem for our system as the FRED code simply overwrites past retrieved data, instead of
accumulating it. This was done under the assumption that FRED was not deleting past data and that
any updates to data would be data revisions. This assumption was incorrect with respect to
`AMERIBOR` where FRED deletes past data.